// assets-ui/src/enums/unitNoteDocTypes.ts
export enum UnitNoteDocTypes {
  NOTICE_OF_CAUTION = 'CAU',
  CONTINUED_NOTE_OF_CAUTION = 'CAUC',
  EXTENSION_TO_NOTICE_OF_CAUTION = 'CAUE',
  NOTE_CANCELLATION = 'NCAN',
  CONFIDENTIAL_NOTE = 'NCON',
  PUBLIC_NOTE = 'NPUB',
  RESTRAINING_ORDER = 'REST',
  NOTICE_OF_TAX_SALE = 'TAXN',
  DECAL_REPLACEMENT = 'REG_102',
  TRANSPORT_PERMIT = 'REG_103',
  AMEND_PERMIT = 'AMEND_PERMIT',
  TRANSPORT_PERMIT_EXTENSION = 'REG_103E',
  NON_RESIDENTIAL_EXEMPTION = 'EXNR',
  RESIDENTIAL_EXEMPTION_ORDER = 'EXRS',
  RESCIND_EXEMPTION = 'EXRE',
  REGISTRATION_CORRECTION = 'REGC',
  NOTICE_OF_REDEMPTION = 'NRED'
}

export enum UnitNoteStatusTypes {
  ACTIVE = 'ACTIVE',
  CANCELLED = 'CANCELLED',
  EXPIRED = 'EXPIRED'
}